Refactored aunka and date computation to own file.
The Kílta date is now timezone aware and the epoch time is set to
Sept 5th, 2001, 0H, 0M, 0S, OU, "America/Chicago"

Fixed mastis_bot to use it.

Altered .aunka to emit only the two elements for which it is defined.

Added .date to produce the full Kílta date string for today.