Begin implementation of remembering bot replies.
This code ultimately will allow the user to edit their message that
prompted the bot to initially reply, which will make the bot edit their
own message with the altered results of the edit the user did.
The usual use case is the use types a .m command and wants to remake it
with correctly spelle Kílta and so the bot should redo its rendering to
mastis in the same post it originally did.