Prototype of getting historical message in command_test_history.
This is proof of concept that I can get historical messages from a channel
in discord.

The point of this code is to later teach the mastis_bot to archive message
content of multiple channels into a database in the event of discord going
away, or changing their policy like charging money to see messages older than X
days, or other enshittification.