Basically the archiving works and backs up the requested text channels.
After it has completed the initial ~4 minute download, the archiving is
generally about a second or two to get zero or a few messages.
I've set it to do an archive pass every 4 hours. That shouldn't cause
anyone to complain.

I need to clean up what gets printed where and add an archiving
statistics command to the mastis bot. Also, I'm not entirely sure
how to trigger the thing manually wrt the coroutines arguing.